улучшена механика сохранений (читабельные индексы, создание папки, мессаг при 0 сэйвов)

+++ lib/modules.rb
@@ -22,21 +22,24 @@ module TextBlocks
     newOrLoad = gets.chomp
     case newOrLoad
     when "s"
-      puts "\nEnter name for your character:"
-      name = gets.chomp.capitalize
-      $player = Player.new(name)
-      puts "\nLet's start your adventure, #{$player.name}!\n"
-      sleep(1.5)
-      get_player_action()
+        game_begin
     when "l"
-      load_game()
-      sleep(1.5)
-      get_player_action()
+        load_game()
+        sleep(1.5)
+        get_player_action() if $loaded
     end
   end
 end
 
 module Action
+    def game_begin
+        puts "\nEnter name for your character:"
+        name = gets.chomp.capitalize
+        $player = Player.new(name)
+        puts "\nLet's start your adventure, #{$player.name}!\n"
+        sleep(1.5)
+        get_player_action()
+    end
   def get_player_action()
     puts "\nWhat do you want to do?\nd - display information about you | i - inspect your items\nm - move to some location | x - exit game"
     user_action = gets.chomp
@@ -138,16 +141,33 @@ module LoadAndSave
 
     end
 
-        def load_game
+    def load_game
+        Dir.mkdir("saves") if !(Dir.exist?("saves"))
         all_saves = Dir.entries("saves").select {|f| !File.directory? f}
-        puts ""
-        all_saves.each_with_index do |record,index|
-        puts "#{record} - #{index}"
+        puts "\nDisplaying all the save files...\n"
+        if all_saves.empty?
+            puts "\nNo save files found :(\n"
+            puts "Do you want to start a new game? y/n"
+            startNew = gets.chomp
+            case startNew
+            when "y"
+                game_begin()
+            when "n"
+                puts "\nOkay then. Farewell!"
+                exit
+            end
+        else
+            all_saves.each_with_index do |record,index|
+                shownIndex = index + 1
+                puts "#{record} - #{shownIndex}"
+            end
+            puts "\nEnter the number of the save file:"
+            shownLoadNumber = gets.chomp
+            loadNumber = shownLoadNumber.to_i - 1
+            File.open("saves/#{all_saves[loadNumber.to_i]}", "r"){|from_file| $player = Marshal.load(from_file)}
+            puts "\nLoaded player: #{$player.name} (exp: #{$player.exp})."
+            File.delete("saves/#{all_saves[loadNumber.to_i]}")
+            $loaded = true
         end
-        puts "\nEnter the number of the save file:"
-        loadNumber = gets.chomp
-        File.open("saves/#{all_saves[loadNumber.to_i]}", "r"){|from_file| $player = Marshal.load(from_file)}
-        puts "\nLoaded player: #{$player.name} (exp: #{$player.exp})."
-        File.delete("saves/#{all_saves[loadNumber.to_i]}")
     end
 end
